mlb-analytics-backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
requests==2.28.1
python-dotenv==0.21.0
pytest==7.2.0
pytest-django==4.5.2
django-cors-headers==4.0.0
celery==5.2.0
redis==4.5.0
pandas==2.0.0
numpy==1.24.0
django-extensions==3.2.0
django-filter==23.2
django-redis==5.2.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2